refactor: Create `codec` module for leaf node.
